Make pre-install a bit more robust.

+++ scripts/before_install.sh
@@ -2,4 +2,7 @@
 
 # Remove the previous repository so we don't get conflicts
 # on trying to copy the new files during deployment.
-rm -rf /home/ec2-user/aqueduct
+if [ -d /home/ec2-user/aqueduct ]; then
+    rm -rf /home/ec2-user/aqueduct
+fi
+mkdir -vp /home/ec2-user/aqueduct
